fix: service stops working eventually There is an issue where the service stops working sporadically after running for a long time.
The only issue identified was in the 'sponsor' route in the controller. It is not clear if this is the issue that causes the service to fail, but it is a problem that could cause the service to not function once all the accounts are irreversibly locked.

The issue in the sponsor route is that we lock a random account, attempt to sponsor a transaction, and then throw an error if the fee is too high. At this point, the account remains locked, which means that the service will begin to fail as soon as we encounter a transaction with a high fee per account and all accounts are locked. This will cause the service to perpetually return a 400 with 'Reached max attempts while locking sponsor address' response thereafter.

The fix is to unlock the account in a `finally` block, regardless of what errors or responses are returned.